MP Fixes Round 1
Patched the worst of the holes, just minor stuff left

+++ Projectiles/VFX/RealityCrack.cs
@@ -80,7 +80,7 @@ namespace tsorcRevamp.Projectiles.VFX
 
                 } while (index < 20);
             }
-            if(filterIndex == null)
+            if(filterIndex == null && Main.netMode != NetmodeID.Server)
             {
                 Projectile.Kill();
                 return;
